requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
python-multipart
email-validator
pydantic
pydantic-settings
psycopg2
langchain 
langchain_core
langchain-openai
matplotlib 
plotly
seaborn
langchain_community
langchain_experimental
jinja2 
pandas

